+++ src/main/java/exception/EmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package exception;
 
 public class EmptyDescriptionException extends Exception {
     public EmptyDescriptionException(String message) {
