+++ src/main/java/exception/InvalidTaskNumberException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package exception;
 
 public class InvalidTaskNumberException extends Exception {
     public InvalidTaskNumberException(String message) {
