+++ src/main/java/exception/TaskListFullException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package exception;
 
 public class TaskListFullException extends Exception {
     public TaskListFullException(String message) {
