+++ src/main/java/exception/UnknownCommandException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package exception;
 
 public class UnknownCommandException extends Exception {
     public UnknownCommandException(String message) {
